PaperDatabase rewrite.
as opposed to individual methods of the PaperDatabase class each creating a
database connection and closing it, the connection is now a PaperDatabase
attribute that is opened and closed via it's __enter__ and __exit__ functions.
i.e., as opposed to using:
>> database = PaperDatabase(path)
>> database.delete_paper()
with delete_paper() using:
>> conn = sqlite3.connect(self.path)
>> with conn:
>>     conn.execute(delete_command)
The new approach looks like:
>> with PaperDatabase(path) as database:
>>     database.delete_paper()
with delete_paper() using:
>> self.conneciton.execute(delete_command)

Other changes:
- moved utils.get_all_rows_from_db() to PaperDatabase.get_all_papers()
- renamed PaperDatabase.query_papers() to PaperDatabase.print_all_papers() and
removed the silent option.
- renamed PaperDatabase.query_papers_contains() to PaperDatabase.query_papers,
which now directly calls PaperDatabase.get_all_papers() and no longer depends on
the old PaperDatabase.query_papers()
- moved alot of code from commands/init.py to PaperDatabase.initalise_database()